Add support for Good General Cauchy Matrices (#2)
add _gc versions of encode/decode that use the 'good cauchy' matrix for the particular k, m and w. If w is not supplied it is calculated as `ceil(math:log2(K+M))`. These should perform better in all cases than the Vandermonde matrixes used in the regular encode/decode.

Note, however, that sometimes w should be higher than the default for ideal performance. This is not done automatically.