Fix builds for Mac OS X and CMake.
Fix builds for Mac OS X hosts using a method that JayKickliter and I derived
while researching this current problem with CMake. In the future, CMake will
probably come up with some reasonable defaults that update CFLAGS properly
on Mac OS X so that it gets the proper `-isysroot <PATH>` that is sorely
needed when invoking `clang` as directly as CMake is doing.